backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.35
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0

